requirements.txt
gradio==4.44.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
Pillow>=9.0.0
numpy>=1.21.0
scikit-learn>=1.3.0
joblib>=1.3.0
